Moving classes to folders to make it little bit cleaner
