+++ .gitignore
@@ -2,3 +2,4 @@
 esdoc
 node_modules
 npm-debug.log
+.DS_Store
