+++ src/core/User.js
@@ -1,0 +1,20 @@
+
+/**
+ * This captures all of the user data associated with the LTI Launch.
+ */
+
+class User{
+
+    /**
+     * @param {number} id
+     */
+    constructor(id) {
+        /**
+         * @type {number}
+         */
+        this.id = id;
+    }
+
+}
+
+module.exports = User;
