--- src/java/Caliper.java
@@ -1,16 +1,0 @@
-
-package org.tsugi;
-
-import java.sql.ResultSet;
-
-/**
- * This is a placeholder for a Caliper API for Tsugi
- */
-public interface Caliper {
-
-    /**
-     * Get the launch associated with this object
-     */
-    public Launch getLaunch();
-
-}
