--- src/java/Link.java
@@ -1,34 +1,0 @@
-
-package org.tsugi;
-
-/**
- * This captures the information around the LTI resource_link
- */
-public interface Link {
-
-    /**
-     * Get the launch associated with this object
-     */
-    public Launch getLaunch();
-
-    /**
-     * The integer primary key for the user within this instance of Tsugi
-     */
-    public Long getId();
-
-    /**
-     * The link title.
-     */
-    public String getTitle();
-
-    /**
-     * The result associated with this link.
-     */
-    public Result getResult();
-
-    /**
-     * The link settings
-     */
-    public Settings getSettings();
-
-}
